+++ src/compas/datastructures/mesh/combinatorics.py
@@ -28,6 +28,15 @@ def mesh_is_connected(mesh):
     -----
     A mesh is connected if for every two vertices a path exists connecting them.
 
+    Examples
+    --------
+    >>> mesh_is_connected(m1)
+    True
+    >>> mesh_is_connected(m2)
+    True
+    >>> mesh_is_connected(m3)
+    False
+
     """
     if not mesh.vertex:
         return False
@@ -43,4 +52,30 @@ def mesh_is_connected(mesh):
 
 if __name__ == "__main__":
 
-    pass
+    import doctest
+    import compas
+    from compas.datastructures import Mesh
+    from compas.datastructures import meshes_join
+    from compas.geometry import translate_points_xy
+    from compas.geometry import subtract_vectors
+
+    m1 = Mesh.from_obj(compas.get('faces.obj'))
+    m2 = m1.copy()
+
+    points = m2.get_vertices_attributes('xyz')
+    x, y, z = zip(*points)
+    xmin = min(x)
+    xmax = max(x)
+    points = translate_points_xy(points, [1.5 * (xmax - xmin), 0, 0])
+    for key, attr in m2.vertices(True):
+        attr['x'] = points[key][0]
+        attr['y'] = points[key][1]
+        attr['z'] = points[key][2]
+
+    m3 = meshes_join([m1, m2])
+
+    doctest.testmod()
+
+
+
+