+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added oriented bounding box for meshes `compas.datastructures.mesh_oriented_bounding_box_numpy`.
+
 ### Changed
 
 ### Removed